Add new page for feature "Add chair" - Add/edit routing
